Persist events table selection after pin/unpin
helps fix #762
